feat: move gc::utils to test_kits crate (#11340)
* feat: move table_test_fixture to test_utils crate

* feat: move table_test_fixture to test_utils crate

* feat: move table_test_fixture to test_kits crate

* feat: move table_test_fixture to test_kits crate

* feat: move gc::utils to test_kits crate

* feat: move gc::utils to test_kits crate

* feat: move gc::utils to test_kits crate

* feat: move gc::utils to test_kits crate
