+++ src/query/service/src/lib.rs
@@ -29,6 +29,7 @@
 #![feature(vec_into_raw_parts)]
 #![feature(associated_type_bounds)]
 #![feature(hash_drain_filter)]
+#![feature(io_error_other)]
 
 extern crate core;
