+++ src/query/service/tests/it/storages/fuse/bloom_index_meta_size.rs
@@ -26,6 +26,7 @@ use common_expression::TableSchemaRefExt;
 use common_storages_fuse::io::TableMetaLocationGenerator;
 use common_storages_fuse::statistics::gen_columns_statistics;
 use common_storages_fuse::FuseStorageFormat;
+use databend_query::test_kits::block_writer::BlockWriter;
 use opendal::Operator;
 use storages_common_cache::InMemoryCacheBuilder;
 use storages_common_cache::InMemoryItemCacheHolder;
@@ -35,8 +36,6 @@ use sysinfo::ProcessExt;
 use sysinfo::System;
 use sysinfo::SystemExt;
 use uuid::Uuid;
-
-use crate::storages::fuse::block_writer::BlockWriter;
 
 // NOTE:
 //
